requirements.txt
PyPDF2~=3.0.1
fastapi~=0.116.1
sympy~=1.14.0
setfit~=1.1.3
openai~=1.102.0
python-dotenv~=1.1.1